requirements.txt
djangorestframework
djangorestframework_jwt
pymongo
colorama
channels
asgi_redis